+++ backend/public_user/src/main/java/com/ics/geomaster/users/models/User.java
@@ -4,7 +4,7 @@ public class User {
   private Integer id;
   private String username;
   private String password;
-  private double balance;
+  private Integer balance;
 
   public User(String username, String password) {
     this.username = username;
@@ -12,7 +12,7 @@ public class User {
     this.balance = 0;
   }
 
-  public User(String username, String password, double balance) {
+  public User(String username, String password, Integer balance) {
     this.username = username;
     this.password = password;
     this.balance = balance;
@@ -28,7 +28,7 @@ public class User {
 
   public double getBalance() { return balance; }
 
-  public void setBalance(double balance) { this.balance = balance; }
+  public void setBalance(Integer balance) { this.balance = balance; }
 
 
   public String toString() {
